refactor: made footer as a common element and linked the upcoming-events
